Make sure we're able to restart the event loop after stopping
The @stopped variable was not reset after stopping the event looop, and
so we were not really able to re-start the event loop after stopping it
once.

The new test is also more accurate as it registers a new handler after
stopping the event loop and waits for its execution in order to pass the
test.